Refactor pv/delta legend as shared
